Functions for reading train-test tfrecord data directly
